ER UI Architecture Redesign: post-close feedback rounds 3-5
Bed graphic restored on bed cards, gendered/age-appropriate waiting
avatars, waiting-lane dev data cleanup, exclusive accordion + infant
age + History tabs + bed lane wrapping fixes. Verified previously via
Playwright (23/23) and manual DOM/screenshot checks per the ER UI
Architecture Redesign tracking doc's rounds 3-5 log entries; committing
now to give the upcoming Legacy Frontend Shell & Dashboard Refresh
branch a clean base.

+++ backend/features/data_management/api.py
@@ -82,7 +82,10 @@ class _DailyPatientBase(BaseModel):
     subject_id: int
     name:        Optional[str]   = None
     gender:      Optional[str]   = None
-    age:         Optional[int]   = None
+    # Float, not int (2026-09-22 fix, matching patient_management/api.py) —
+    # the DB column is already Float; an infant's age needs sub-year
+    # precision, which a strict-int Pydantic field silently made impossible.
+    age:         Optional[float] = None
     arrival_time: Optional[str] = None
     departure_time: Optional[str] = None
     bed_occupation_time: Optional[str] = None
